LayoutTests/inspector/debugger/live-edit-breakpoints-expected.txt

Tests breakpoints are correctly dimmed and restored in JavaScriptSourceFrame during live edit.

Bug 99598
Debugger was enabled.

Running: testEditUndo
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Editing source:
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Undoing source editing:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Committing edited source:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommitFailEditCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Editing source again so that live edit could succeed:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Committing edited source again:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommitFailUndoCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Undoing source editing:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Committing edited source again:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network, project id: file://
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditAnonymousUndo
Set timer for test function.
Script execution paused.
Script execution resumed.
Showing script source:
Setting breakpoint:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Editing source:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Undoing source editing:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Finally removing breakpoint:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditDynamicAnonymousUndo
Set timer for test function.
Script execution paused.
Script execution resumed.
Showing script source:
Setting breakpoint:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Editing source:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Undoing source editing:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project type: debugger, project id: debugger:
    Dumping breakpoints
        url: , lineNumber: 2, project type: debugger, project id: debugger:
Finally removing breakpoint:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints
Debugger was disabled.
